Jiras RE-1481 & RE-1482 corrected the header issue on the teams scorecard & updated the fields for SLAs to no longer be decimal fields they are not integerfields -- they were no used previously so the issue wasn't detected until now.

+++ scorecard/apps/teams/models.py
@@ -359,10 +359,14 @@ class RequirementMetrics(BaseMetrics):
     # This is now system_miss as it's pulling from Personal now
     slas_missed = models.DecimalField(max_digits=3, decimal_places=2, default=0,
                                       verbose_name='SLAs Missed')
-    system_met = models.DecimalField(max_digits=3, decimal_places=2, default=0, verbose_name='System SLA Met')
-    system_miss = models.DecimalField(max_digits=3, decimal_places=2, default=0, verbose_name='System SLA Miss')
-    actual_met = models.DecimalField(max_digits=3, decimal_places=2, default=0, verbose_name='Actual SLA Met')
-    actual_miss = models.DecimalField(max_digits=3, decimal_places=2, default=0, verbose_name='Actual SLA Miss')
+    system_met = models.PositiveIntegerField(default=0, verbose_name='System SLA Met')
+    system_miss = models.PositiveIntegerField(default=0, verbose_name='System SLA Miss')
+    actual_met = models.PositiveIntegerField(default=0, verbose_name='Actual SLA Met')
+    actual_miss = models.PositiveIntegerField(default=0, verbose_name='Actual SLA Miss')
+    # system_met = models.DecimalField(max_digits=3, decimal_places=2, default=0, verbose_name='System SLA Met')
+    # system_miss = models.DecimalField(max_digits=3, decimal_places=2, default=0, verbose_name='System SLA Miss')
+    # actual_met = models.DecimalField(max_digits=3, decimal_places=2, default=0, verbose_name='Actual SLA Met')
+    # actual_miss = models.DecimalField(max_digits=3, decimal_places=2, default=0, verbose_name='Actual SLA Miss')
 
     # Optimization
     optimization_time = models.DecimalField(max_digits=10, decimal_places=2, default=0,
